Refactor: Split Missions.tsx into smaller components
Split the Missions page into smaller, more manageable components to improve readability and maintainability.
